Stop checking project URL.
This is a hassle to keep up to date.

It turns out to be low-value for automation, since it's better to
check code as shipped on npm, not as stored on github. For that, we just
need to check that a DT package name has an associated package on npm.
We don't need to make sure that the github url is present too.

I don't think it is that useful to people either, since they can find
the homepage of the original package on npmjs.com.